Batch: Move SGraph and SFilter into separate source files
